fix: vt를 이용해서 크롤링을 하면서 다음 url이 제대로 추가가 안되는 문제 해결
